# docs/openapi/retail-ec.yaml
openapi: 3.0.3
info:
  title: Retail/EC API
  description: |
    Multi-tenant inventory platform — Retail/EC service (1st of the four
    business contexts).

    Scope: order acceptance and lookup. Place an order and the service emits
    `retail.order.placed.v1` for downstream services (Inventory Core,
    Notification, Integration Hub) to consume.

    Bridge multi-tenancy: each tenant has its own schema.
    Errors follow RFC 7807; `errorCode` is the i18n key.
  version: 1.0.0

servers:
  - url: https://api.example.com/retail-ec
    description: Production
  - url: http://localhost:8085
    description: Local dev

tags:
  - name: orders
    description: Order placement and lookup

security:
  - bearerJwt: []

paths:
  /v1/orders:
    post:
      tags: [orders]
      operationId: placeOrder
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Order code already exists in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /v1/orders/{orderId}:
    parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      tags: [orders]
      operationId: getOrder
      summary: Get order by id
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/OrderNotFound'

  /v1/orders/{orderId}/ship:
    parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    post:
      tags: [orders]
      operationId: shipOrder
      summary: Confirm shipment of a placed order, emitting retail.order.shipped.v1
      responses:
        '200':
          description: Order shipped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          description: Order is not in a state that allows shipment
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /v1/orders/{orderId}/cancel:
    parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    post:
      tags: [orders]
      operationId: cancelOrder
      summary: Cancel a placed order, emitting retail.order.cancelled.v1
      description: |
        業務側起因のキャンセル。 Inventory Core が listener で受け、 reserved を release する。
        既に CANCELLED の場合は冪等で 200 を返す(イベント発行なし)。 SHIPPED からの cancel は 409。
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          description: Order is not in a state that allows cancellation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  schemas:
    PlaceOrderRequest:
      type: object
      required: [code, customerEmail, currency, items]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9][A-Z0-9-]{2,63}$'
          description: Tenant-unique order code
          example: ORD-2026-0001
        customerEmail:
          type: string
          maxLength: 256
          example: alice@example.com
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: JPY
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderLine'

    OrderLine:
      type: object
      required: [skuCode, locationId, quantity, unitPrice]
      properties:
        skuCode:
          type: string
          maxLength: 64
          example: SKU-COCA-COLA-500ML
        locationId:
          type: string
          maxLength: 64
          description: |
            Source location for this line. Inventory Core resolves the inventory
            record by (skuCode, locationId) when consuming `retail.order.placed.v1`.
          example: LOC-TOKYO-A
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000000
        unitPrice:
          type: number
          format: double
          minimum: 0

    OrderResponse:
      type: object
      required: [id, code, status, currency, totalAmount, items, version]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [PLACED, SHIPPED, CANCELLED]
        currency:
          type: string
        totalAmount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        version:
          type: integer
          format: int64

    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        errorCode: { type: string }
        traceId: { type: string }

  responses:
    Unauthorized:
      description: Not authenticated, JWT invalid/expired
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    OrderNotFound:
      description: Order does not exist in this tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT